Move DeleteRoomButton
