+++ src/app/(admin)/admin/settings/rooms/[id]/page.tsx
@@ -5,7 +5,7 @@ import React from "react";
 import Link from "next/link";
 import { api } from "~/trpc/react";
 import { InfoCard } from "~/app/_components/admin/InfoCard";
-import DeleteRoomButton from "~/app/_components/admin/DeleteRoomButton";
+import DeleteRoomButton from "~/app/(admin)/admin/settings/rooms/[id]/DeleteRoomButton";
 
 
 const ShowRoom = ({  params: { id }  }: { params: { id: string } }) => {
